enable decompiler on windows
I'm not sure why the decompiler was disabled on windows, it works just
fine.

The only reason I can think of is that it requires nokogiri which is a C
extension, but the dep on nokogiri isn't exempted, so windows users who
did `gem install twee2` were getting it anyway.

Additionally, there are precompiled versions of nokogiri available for
windows now, so you don't need a C compiler.

Furthermore, RubyInstaller can install the C compilers now.